requirements.txt
asgiref==3.5.0
Babel==2.9.1
backports.zoneinfo==0.2.1
Django==4.0.2
django-money==2.1.1
djangorestframework==3.13.1
py-moneyed==1.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
